+++ BuyRateRefresher.cs
@@ -1,4 +1,5 @@
 ﻿using UnityEngine;
+using BuyRateSettings.Configuration;
 
 namespace BuyRateSettings;
 
@@ -8,30 +9,32 @@ public static class BuyRateRefresher
     {
         // Variables
         float price = StartOfRound.Instance.companyBuyingRate;
+        double rateSeed = StartOfRound.Instance.randomMapSeed * 3 + 99; // +99 because starting seed is 0, which causes a jackpot on day 1 // *3 just so the rateSeed is diff from the mapSeed
+        double rateSeedRemainder = rateSeed % 100 / 100;
 
-        bool minMaxToggle = BuyRateModifier.minMaxToggle.Value;
-        float minRate = BuyRateModifier.minRate.Value;
-        float maxRate = BuyRateModifier.maxRate.Value;
+        bool minMaxToggle = Config.Instance.minMaxToggle;
+        float minRate = Config.Instance.minRate;
+        float maxRate = Config.Instance.maxRate;
 
-        bool randomRateToggle = BuyRateModifier.randomRateToggle.Value;
+        bool randomRateToggle = Config.Instance.randomRateToggle;
 
-        bool lastDayToggle = BuyRateModifier.lastDayToggle.Value;
-        float lastDayRangeChance = BuyRateModifier.lastDayRangeChance.Value;
-        float lastDayMinRate = BuyRateModifier.lastDayMinRate.Value;
-        float lastDayMaxRate = BuyRateModifier.lastDayMaxRate.Value;
+        bool lastDayToggle = Config.Instance.lastDayToggle;
+        float lastDayRangeChance = Config.Instance.lastDayRangeChance;
+        float lastDayMinRate = Config.Instance.lastDayMinRate;
+        float lastDayMaxRate = Config.Instance.lastDayMaxRate;
         float daysUntilDeadline = TimeOfDay.Instance.daysUntilDeadline;
 
-        bool jackpotToggle = BuyRateModifier.jackpotToggle.Value;
-        bool jackpotToggleLD = BuyRateModifier.jackpotToggleLD.Value;
-        double jackpotChance = BuyRateModifier.jackpotChance.Value;
-        float jackpotMinRate = BuyRateModifier.jackpotMinRate.Value;
-        float jackpotMaxRate = BuyRateModifier.jackpotMaxRate.Value;
+        bool jackpotToggle = Config.Instance.jackpotToggle;
+        bool jackpotToggleLD = Config.Instance.jackpotToggleLD;
+        double jackpotChance = Config.Instance.jackpotChance;
+        float jackpotMinRate = Config.Instance.jackpotMinRate;
+        float jackpotMaxRate = Config.Instance.jackpotMaxRate;
         bool jackpotHit = false;
 
-        bool jackpotAlertToggle = BuyRateModifier.jackpotAlertToggle.Value;
-        bool buyRateAlertToggle = BuyRateModifier.buyRateAlertToggle.Value;
+        bool jackpotAlertToggle = Config.Default.jackpotAlertToggle; // Client sided
+        bool buyRateAlertToggle = Config.Default.buyRateAlertToggle; // Client sided
 
-        float alertDelaySeconds = BuyRateModifier.alertDelaySeconds.Value;
+        float alertDelaySeconds = Config.Instance.alertDelaySeconds;
         float rateDelayTimeSeconds = 3; // Hardcoded to set the rate a 2nd time 3s later to prevent being overwritten by other mods
 
 
@@ -47,9 +50,16 @@ public static class BuyRateRefresher
         // Min/Max
         // Vanilla
 
+        BuyRateModifier.mls.LogInfo($"Days left: {TimeOfDay.Instance.daysUntilDeadline}");
+        BuyRateModifier.mls.LogInfo($"Initial buying rate (pre-calculation): {StartOfRound.Instance.companyBuyingRate}");
+        BuyRateModifier.mls.LogInfo($"Jackpot chance: {jackpotChance}");
+        BuyRateModifier.mls.LogInfo($"Map seed: {StartOfRound.Instance.randomMapSeed}");
+        BuyRateModifier.mls.LogInfo($"Rate seed: {rateSeed}");
+        BuyRateModifier.mls.LogInfo($"Rate seed remainder: {rateSeedRemainder}");
+
 
         // Check for jackpot on last day and roll
-        if (jackpotToggle && DateTime.Now.Millisecond % 100 / 100.0 <= jackpotChance && jackpotToggleLD && daysUntilDeadline <= 0)
+        if (jackpotToggle && rateSeedRemainder <= jackpotChance && jackpotToggleLD && daysUntilDeadline <= 0)
         {
             // Last day jackpot range
             if (jackpotMinRate != jackpotMaxRate)
@@ -57,7 +67,7 @@ public static class BuyRateRefresher
                 price = Next() * (jackpotMaxRate - jackpotMinRate) + jackpotMinRate;
                 jackpotHit = true;
 
-                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - LAST DAY ONLY (RANGED) (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - LAST DAY ONLY (RANGED) (unrounded rate: {price})");
             }
 
             // Last day jackpot no range
@@ -66,12 +76,12 @@ public static class BuyRateRefresher
                 price = jackpotMinRate;
                 jackpotHit = true;
 
-                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - LAST DAY ONLY (NOT RANGED) (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - LAST DAY ONLY (NOT RANGED) (unrounded rate: {price})");
             }
         }
 
         // Jackpot on any day and roll
-        else if (jackpotToggle && DateTime.Now.Millisecond % 100 / 100.0 <= jackpotChance && jackpotToggleLD == false && daysUntilDeadline > 0)
+        else if (jackpotToggle && rateSeedRemainder <= jackpotChance && jackpotToggleLD == false)
         {
             // Any day jackpot ranged
             if (jackpotMinRate != jackpotMaxRate)
@@ -79,7 +89,7 @@ public static class BuyRateRefresher
                 price = Next() * (jackpotMaxRate - jackpotMinRate) + jackpotMinRate;
                 jackpotHit = true;
 
-                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - ANY DAY (RANGED) (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - ANY DAY (RANGED) (unrounded rate: {price})");
             }
 
             // Any day jackpot no range
@@ -88,7 +98,7 @@ public static class BuyRateRefresher
                 price = jackpotMinRate;
                 jackpotHit = true;
 
-                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - ANY DAY (NOT RANGED) (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"HIT THE JACKPOT - ANY DAY (NOT RANGED) (unrounded rate: {price})");
             }
         }
 
@@ -96,25 +106,25 @@ public static class BuyRateRefresher
         else if (lastDayToggle && daysUntilDeadline <= 0)
         {
             // Last day random range hit
-            if (lastDayMinRate != lastDayMaxRate && DateTime.Now.Millisecond % 100 / 100.0 <= lastDayRangeChance)
+            if (lastDayMinRate != lastDayMaxRate && rateSeedRemainder <= lastDayRangeChance)
             {
                 price = Next() * (lastDayMaxRate - lastDayMinRate) + lastDayMinRate;
 
-                BuyRateModifier.mls.LogInfo($"Last day rate (ranged-hit) picked (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"Last day rate (ranged-hit) picked (unrounded rate: {price})");
             }
 
             // Last day random range not hit (default to 100%)
-            else if (lastDayMinRate != lastDayMaxRate && DateTime.Now.Millisecond % 100 / 100.0 > lastDayRangeChance)
+            else if (lastDayMinRate != lastDayMaxRate && rateSeedRemainder > lastDayRangeChance)
             {
                 price = 1f;
-                BuyRateModifier.mls.LogInfo($"Last day rate (ranged-nohit) picked (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"Last day rate (ranged-nohit) picked (unrounded rate: {price})");
             }
 
             // Last day no range
             else
             {
                 price = lastDayMinRate;
-                BuyRateModifier.mls.LogInfo($"Last day rate (not ranged) picked (unrounded rate: {price}");
+                BuyRateModifier.mls.LogInfo($"Last day rate (not ranged) picked (unrounded rate: {price})");
             }
         }
 
@@ -123,28 +133,28 @@ public static class BuyRateRefresher
         {
             price = Next() * (maxRate - minRate) + minRate;
 
-            BuyRateModifier.mls.LogInfo($"Random Rate picked (unrounded rate: {price}");
+            BuyRateModifier.mls.LogInfo($"Random Rate picked (unrounded rate: {price})");
         }
 
         // Set minimum rate
         else if (minMaxToggle && StartOfRound.Instance.companyBuyingRate <= minRate)
         {
             price = minRate;
-            BuyRateModifier.mls.LogInfo($"Min rate picked (unrounded rate: {price}");
+            BuyRateModifier.mls.LogInfo($"Min rate picked (unrounded rate: {price})");
         }
 
         // Set maximum rate
         else if (minMaxToggle && StartOfRound.Instance.companyBuyingRate >= maxRate)
         {
             price = maxRate;
-            BuyRateModifier.mls.LogInfo($"Max rate picked (unrounded rate: {price}");
+            BuyRateModifier.mls.LogInfo($"Max rate picked (unrounded rate: {price})");
         }
 
         // Default definition (vanilla)
         else
         {
             price = StartOfRound.Instance.companyBuyingRate;
-            BuyRateModifier.mls.LogInfo($"Vanilla rate picked (unrounded rate: {price}");
+            BuyRateModifier.mls.LogInfo($"Vanilla rate picked (unrounded rate: {price})");
         }
 
 
@@ -156,7 +166,7 @@ public static class BuyRateRefresher
         // Round price for cleaner text
         int priceRounded = (int)Math.Round(price * 100);
 
-        // Set buy rate
+        // Set buy rate immediately & start coroutine for delayed set
         StartOfRound.Instance.companyBuyingRate = price;
         BuyRateModifier.mls.LogInfo("Set Company buy rate to: " + priceRounded + "%");
         TimeOfDay.Instance.StartCoroutine(BuyRateSetter(rateDelayTimeSeconds, price, priceRounded)); // Reassigns the buy rate a 2nd time in case a new deadline forced a vanilla calculation
@@ -190,7 +200,7 @@ public static class BuyRateRefresher
     {
         yield return new WaitForSecondsRealtime(alertDelayTimeSeconds);
         HUDManager.Instance.DisplayTip("New Scrap Rate", "\n* Buying rates have changed to " + buyRateRounded + "%", false, false, "LC_JackpotTip2");
-        BuyRateModifier.mls.LogInfo("Sent normal alert");
+        BuyRateModifier.mls.LogInfo($"Sent normal alert. Rate: {buyRateRounded}");
     }
 
     private static IEnumerator BuyRateAlertJackpot(float alertDelayTimeSeconds, int buyRateRounded)
@@ -198,6 +208,6 @@ public static class BuyRateRefresher
         yield return new WaitForSecondsRealtime(alertDelayTimeSeconds);
         HUDManager.Instance.DisplayTip("<color=#ffc526>SCRAP EMERGENCY</color>", "<color=#fcbf17>\n* Buying rates have soared to " + buyRateRounded + "%</color>", true, false, "LC_JackpotTip1");
         HUDManager.Instance.UIAudio.PlayOneShot(HUDManager.Instance.globalNotificationSFX);
-        BuyRateModifier.mls.LogInfo("Sent jackpot alert");
+        BuyRateModifier.mls.LogInfo($"Sent jackpot alert. Rate: {buyRateRounded}");
     }
 }